agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strands-agent"
version = "1.0.0"
description = "A Strands AI agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "strands-agents[otel]>=1.19.0",
    "strands-agents-tools>=0.2.17",
    "bedrock-agentcore>=1.1.1",
    "aws-opentelemetry-distro>=0.14.0",
    "boto3>=1.42.9"
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.9",
    "mypy>=1.19.0",
    "black>=25.12.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
mypy_path = "src"
explicit_package_bases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
